+++ kanavka/settings.py
@@ -29,7 +29,7 @@ SECRET_KEY = os.environ.get('DJANGO_SECRET_KEY', 'aaaa(bbbb(+cccc+)1111)2222')
 # DEBUG = True
 DEBUG = bool(os.environ.get('DJANGO_DEBUG', True))
 
-ALLOWED_HOSTS = ['tehkontrol-spup.herokuapp.com']
+ALLOWED_HOSTS = ['tehkontrol-spup.herokuapp.com', '127.0.0.1']
 
 
 # Application definition
