+++ templates/base.html
@@ -11,7 +11,7 @@
         {% block top %}{% endblock top %}
     </div>
     <div id="menu">
-        <a href="/kanavka.html">
+        <a href="{% url 'kanavka' %}">
             <button class="but_1">Канавка</button>
         </a>
         <a href="/rast.html">
